@@ -29,7 +29,7 @@
     "url": "https://github.com/sagalbot/vue-select.git"
   },
   "peerDependencies": {
-    "vue": "2.x"
+    "vue": "3.x"
   },
   "resolutions": {
     "ajv": "6.8.1"
@@ -42,7 +42,8 @@
     "@babel/runtime": "^7.4.2",
     "@semantic-release/git": "^9.0.0",
     "@semantic-release/github": "^7.0.4",
-    "@vue/test-utils": "^1.2.2",
+    "@vue/compiler-sfc": "^3.2.20",
+    "@vue/test-utils": "^2.0.0-rc.16",
     "autoprefixer": "^9.4.7",
     "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.3",
@@ -74,13 +75,12 @@
     "semantic-release": "^17.0.4",
     "terser-webpack-plugin": "^1.2.3",
     "url-loader": "^1.1.2",
-    "vue": "^2.6.10",
+    "vue": "^3.2.20",
     "vue-html-loader": "^1.2.4",
-    "vue-jest": "^3.0.4",
-    "vue-loader": "^15.7.0",
+    "vue-jest": "5.0.0-alpha.8",
+    "vue-loader": "^16.8.1",
     "vue-server-renderer": "^2.6.10",
     "vue-style-loader": "^4.1.2",
-    "vue-template-compiler": "^2.6.10",
     "webpack": "^4.29.6",
     "webpack-cli": "^3.3.0",
     "webpack-dev-server": "^3.2.1",
@@ -129,7 +129,7 @@
       {
         "path": "./dist/vue-select.js",
         "compression": "none",
-        "maxSize": "21 KB"
+        "maxSize": "23 KB"
       },
       {
         "path": "./dist/vue-select.css",
