@@ -4,7 +4,7 @@ import VueSelect from '../../src/components/Select'
 describe('Filtering Options', () => {
   it("should update the search value when the input element receives the 'input' event", () => {
     const Select = shallowMount(VueSelect, {
-      propsData: { options: ['foo', 'bar', 'baz'] },
+      props: { options: ['foo', 'bar', 'baz'] },
     })
 
     const input = Select.find('.vs__search')
@@ -15,7 +15,7 @@ describe('Filtering Options', () => {
 
   it('should filter an array of strings', () => {
     const Select = shallowMount(VueSelect, {
-      propsData: { options: ['foo', 'bar', 'baz'] },
+      props: { options: ['foo', 'bar', 'baz'] },
     })
     Select.vm.search = 'ba'
     expect(Select.vm.filteredOptions).toEqual(['bar', 'baz'])
@@ -23,7 +23,7 @@ describe('Filtering Options', () => {
 
   it('should not filter the array of strings if filterable is false', () => {
     const Select = shallowMount(VueSelect, {
-      propsData: { options: ['foo', 'bar', 'baz'], filterable: false },
+      props: { options: ['foo', 'bar', 'baz'], filterable: false },
     })
     Select.vm.search = 'ba'
     expect(Select.vm.filteredOptions).toEqual(['foo', 'bar', 'baz'])
@@ -31,7 +31,7 @@ describe('Filtering Options', () => {
 
   it('should filter without case-sensitivity', () => {
     const Select = shallowMount(VueSelect, {
-      propsData: { options: ['Foo', 'Bar', 'Baz'] },
+      props: { options: ['Foo', 'Bar', 'Baz'] },
     })
     Select.vm.search = 'ba'
     expect(Select.vm.filteredOptions).toEqual(['Bar', 'Baz'])
@@ -39,7 +39,7 @@ describe('Filtering Options', () => {
 
   it('can filter an array of objects based on the objects label key', () => {
     const Select = shallowMount(VueSelect, {
-      propsData: {
+      props: {
         options: [{ label: 'Foo' }, { label: 'Bar' }, { label: 'Baz' }],
       },
     })
@@ -52,7 +52,7 @@ describe('Filtering Options', () => {
 
   it('can determine if a given option should match the current search text', () => {
     const Select = shallowMount(VueSelect, {
-      propsData: {
+      props: {
         options: [{ label: 'Aoo' }, { label: 'Bar' }, { label: 'Baz' }],
         filterBy: (option, label, search) =>
           label.match(new RegExp('^' + search, 'i')),
@@ -65,7 +65,7 @@ describe('Filtering Options', () => {
 
   it('can use a custom filtering method', () => {
     const Select = shallowMount(VueSelect, {
-      propsData: {
+      props: {
         options: ['foo', 'bar', 'baz'],
         filterBy: (option, label) => label.includes('o'),
       },
@@ -76,7 +76,7 @@ describe('Filtering Options', () => {
 
   it('can filter arrays of numbers', () => {
     const Select = shallowMount(VueSelect, {
-      propsData: {
+      props: {
         options: [1, 5, 10],
       },
     })
