@@ -4,7 +4,7 @@ import {
   selectTag,
   selectWithProps,
 } from '../helpers'
-import Select from '../../src/components/Select'
+import VueSelect from '../../src/components/Select'
 
 describe('When Tagging Is Enabled', () => {
   it('can determine if a given option string already exists', () => {
@@ -40,7 +40,7 @@ describe('When Tagging Is Enabled', () => {
     const Select = selectWithProps({
       taggable: true,
       multiple: true,
-      value: ['one'],
+      modelValue: ['one'],
       options: ['one', 'two'],
     })
 
@@ -78,7 +78,7 @@ describe('When Tagging Is Enabled', () => {
       pushTags: true,
       taggable: true,
       multiple: true,
-      value: ['one'],
+      modelValue: ['one'],
       options: ['one', 'two'],
     })
 
@@ -107,7 +107,7 @@ describe('When Tagging Is Enabled', () => {
       pushTags: true,
       taggable: true,
       multiple: true,
-      value: ['one'],
+      modelValue: ['one'],
       options: ['one', 'two'],
     })
 
@@ -122,7 +122,7 @@ describe('When Tagging Is Enabled', () => {
       pushTags: false,
       taggable: true,
       multiple: true,
-      value: ['one'],
+      modelValue: ['one'],
       options: ['one', 'two'],
     })
 
@@ -136,7 +136,7 @@ describe('When Tagging Is Enabled', () => {
       pushTags: false,
       taggable: true,
       multiple: true,
-      value: ['one'],
+      modelValue: ['one'],
       options: ['one', 'two'],
     })
 
@@ -191,7 +191,7 @@ describe('When Tagging Is Enabled', () => {
     const Select = selectWithProps({
       taggable: true,
       multiple: true,
-      value: [{ label: 'one' }],
+      modelValue: [{ label: 'one' }],
       options: [{ label: 'one' }],
     })
 
@@ -204,7 +204,7 @@ describe('When Tagging Is Enabled', () => {
       taggable: true,
       multiple: true,
       filterable: false,
-      value: [{ label: 'one' }],
+      modelValue: [{ label: 'one' }],
       options: [{ label: 'one' }],
     })
 
@@ -228,12 +228,12 @@ describe('When Tagging Is Enabled', () => {
   })
 
   it('should not allow duplicate tags when using object options', async () => {
+    const spy = jest.spyOn(VueSelect.methods, 'select')
     const Select = selectWithProps({
       taggable: true,
       multiple: true,
       options: [{ label: 'two' }],
     })
-    const spy = jest.spyOn(Select.vm, 'select')
 
     await selectTag(Select, 'one')
     expect(Select.vm.selectedValue).toEqual([{ label: 'one' }])
@@ -252,9 +252,8 @@ describe('When Tagging Is Enabled', () => {
     })
 
     Select.vm.typeAheadPointer = 0
-    Select.findComponent({ ref: 'search' }).trigger('keydown.tab')
+    await Select.get('input').trigger('keydown.tab')
 
-    await Select.vm.$nextTick()
     expect(Select.vm.selectedValue).toEqual(['one'])
   })
 })
