@@ -7,7 +7,7 @@ import Vue from 'vue'
 describe('Moving the Typeahead Pointer', () => {
   it('should set the pointer to zero when the filteredOptions watcher is called', async () => {
     const Select = shallowMount(VueSelect, {
-      propsData: { options: ['one', 'two', 'three'] },
+      props: { options: ['one', 'two', 'three'] },
       sync: false,
     })
 
@@ -22,7 +22,7 @@ describe('Moving the Typeahead Pointer', () => {
 
     Select.vm.typeAheadPointer = 1
 
-    Select.findComponent({ ref: 'search' }).trigger('keydown.up')
+    Select.get('input').trigger('keydown.up')
 
     expect(Select.vm.typeAheadPointer).toEqual(0)
   })
@@ -32,7 +32,7 @@ describe('Moving the Typeahead Pointer', () => {
 
     Select.vm.typeAheadPointer = 1
 
-    Select.findComponent({ ref: 'search' }).trigger('keydown.down')
+    Select.get('input').trigger('keydown.down')
 
     expect(Select.vm.typeAheadPointer).toEqual(2)
   })
